transaksi

+++ application/models/Produk_model.php
@@ -66,35 +66,42 @@ class Produk_model extends CI_Model
 	
 	function process()
 	{
-		$invoice = array(
-			'date' => date('Y-m-d H:i:s'),
-			'due_date' => date('Y-m-d H:i:s', mktime(date('H'),date('i'),date('s'),date('m'),date('d') + 1,date('Y'))),
-			'status' => 'UNPAID',
-			'nama' => $this->input->post('nama',true),
-			'nope' => $this->input->post('nope',true),
-			'email' => $this->input->post('email',true),
-			'alamat' => $this->input->post('alamat',true),
-			'kota_tujuan' => $this->input->post('kota',true),
-			'kurir' => $this->input->post('kurir',true),
-			);
-		$this->db->insert('invoices', $invoice);
-		$invoice_id = $this->db->insert_id();
-		
-		foreach($this->cart->contents() as $item)
-		{
-			$data = array(
-				'invoice_id' => $invoice_id,
-				'product_id' => $item['id'],
-				'product_name' => $item['name'],
-				'qty' 		=> $item['qty'],
-				'price' 	=> $item['price'],
-				);
-			$this->db->insert('orders',$data);
+		if($this->cart->contents() == null){
+			redirect('welcome');
 		}
-		
-		$this->cart->destroy();
-		$this->load->view('order_success',$data);
-		return TRUE;
+		else {
+			$paket = $this->input->post('paket',true);
+			$invoice = array(
+				'date' => date('Y-m-d H:i:s'),
+				'due_date' => date('Y-m-d H:i:s', mktime(date('H'),date('i'),date('s'),date('m'),date('d') + 1,date('Y'))),
+				'status' => 'UNPAID',
+				'nama' => $this->input->post('nama',true),
+				'nope' => $this->input->post('nope',true),
+				'email' => $this->input->post('email',true),
+				'alamat' => $this->input->post('alamat',true),
+				'kota_tujuan' => $this->input->post('kota',true),
+				'kurir' => strtoupper($this->input->post('kurir',true) . ' ' . explode("-", $paket)[0]),
+				'total_bayar' => $this->input->post('total',true),
+			);
+			$this->db->insert('invoices', $invoice);
+			$invoice_id = $this->db->insert_id();
+			
+			foreach($this->cart->contents() as $item)
+			{
+				$data = array(
+					'invoice_id' => $invoice_id,
+					'product_id' => $item['id'],
+					'product_name' => $item['name'],
+					'qty' 		=> $item['qty'],
+					'price' 	=> $item['price'],
+				);
+				$this->db->insert('orders',$data);
+			}
+			
+			$this->cart->destroy();
+			$this->load->view('order_success',$data);
+			return TRUE;
+		}
 	}
 	
 	function all_invoices()
@@ -136,7 +143,9 @@ class Produk_model extends CI_Model
 	{
 		// $this->db->where('invoice_id',$invoice_id);
 		// return $this->db->get('orders')->result();
-		return $this->db->query("SELECT * FROM orders o JOIN invoices i ON i.id = o.invoice_id
+		return $this->db->query("SELECT k.*, o.product_id, o.product_name, o.qty, o.weight, o.price, o.options, i.*
+			FROM konfirmasi k JOIN orders o ON o.invoice_id = k.invoice_id
+			JOIN invoices i ON i.id = o.invoice_id 
 			WHERE o.invoice_id = '$invoice_id'")->result();
 	}
